backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Supabase and database
supabase==2.16.0
httpx==0.28.1
requests==2.32.3

# Token encryption
cryptography==41.0.7

# OpenAI for sales opportunity analysis
openai==1.3.7

# Pydantic for data models
pydantic==2.5.0

# Environment utilities
python-dotenv==1.0.0

# Logging and error handling
structlog==23.2.0

# CORS
python-cors==1.7.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1 